refactor: 소켓 관련 설정, 컴포넌트, 컨트롤러 다른 폴더로 이동
